+++ src/database/build_database.sql
@@ -11,7 +11,7 @@ CREATE TABLE users (
 );
 
 INSERT INTO users (id, given_names, family_name, birth_date, phone_number, admin) VALUES ('xFv5nFhg74HjsolcjDeDDhhbvgcf08G9f4Xf1', 'Nori', 'Denes', '1/1/2000', 07123456789, false);
-INSERT INTO users (id, given_names, family_name, birth_date, phone_number, admin) VALUES ('Re5xYlrYsu8ctaxPN2g7zZYGHSKOHotA6UYABPmadPvMaPs3ssziv0iEBd75zfi2', 'MARKO', 'SUSTARSIC', '1988-02-21', 447811110722, false);
+INSERT INTO users (id, given_names, family_name, birth_date, phone_number, admin) VALUES ('Re5xYlrYsu8ctaxPN2g7zZYGHSKOHotA6UYABPmadPvMaPs3ssziv0iEBd75zfi2', 'MARKO', 'SUSTARSIC', '1988-02-21', 447811110722, true);
 
 SELECT * FROM users;
