+++ src/routes/dashboard.js
@@ -1,14 +1,6 @@
-const fs = require('fs');
-const path = require('path');
-const YotiClient = require('yoti-node-sdk');
-const userTable = require('../database/tables/users');
-const userHelper = require('../helpers/userHelper');
-const retrieveData = require('./../database/tables/retrieveRequests.js')
+const authMiddleware = require('../helpers/authMiddleware.js')
+const retrieveData = require('../database/tables/retrieveRequests.js')
 
-const CLIENT_SDK_ID = "8a4dcb2a-9ed6-4d44-9a55-12b581bb5e64";
-const PEM = fs.readFileSync(path.join(__dirname, '../../keys/help-access-security.pem'));
-const yotiClient = new YotiClient(CLIENT_SDK_ID, PEM);
-const loginError = "There was a problem accessing your Yoti, please log in again. If this problem persists, contact our technical team.";
 const databaseError = "There was a problem retrieving data, please try again. If this problem persists, contact our technical team.";
 const authenticationError = "You are not authenticated to access the data. If you are a Centrepoint administrator, contact our technical team."
 
@@ -19,46 +11,24 @@ module.exports = {
     auth: {
       strategy: 'base'
     },
+    pre: [
+        { method: authMiddleware, assign: 'data' }
+    ],
     handler: (req, reply) => {
-      let token = req.auth.credentials.auth;
-      if(!token) {
-        console.log("No token provided");
+      if (! req.pre.data[0].admin) {
         return reply.view('error', {
-          error : loginError
+          error : authenticationError
         });
       }
-      yotiClient
-      .getActivityDetails(token)
-      .then((activityDetails) => {
-        let user = userHelper.getUser(activityDetails);
-        userTable.queryUserId(user, function (err, data) {
-          if (err) {
-            console.log("Error adding user: ", err);
-            return reply.view('error', {
-              error : databaseError
-            });
-          }
-          if (! data[0].admin) {
-            return reply.view('error', {
-              error : authenticationError
-            });
-          }
-          retrieveData.retrieve(function (err, data) {
-            if (err) {
-              console.log("Error retrieving requests: ", err);
-              return reply.view('error', {
-                error: databaseError
-              });
-            }
-            reply.view('dashboard', { requests: data.rows });
+      retrieveData.retrieve(function (err, data) {
+        if (err) {
+          console.log("Error retrieving requests: ", err);
+          return reply.view('error', {
+            error: databaseError
           });
-        });
-      }).catch((err) => {
-        console.error(err);
-        reply.view('error', {
-          error : yotiError
-        });
-      })
+        }
+        reply.view('dashboard', { requests: data.rows });
+      });
     }
   }
 };
